Reviewer 2
I am having trouble running the app as it is intended. I have tried running it on two different emulators. The first emulator that I tried running was the Google Nexus phone on genymotion, and I am prompted with an error while trying to open the app which says "Uncaught error: Requested Module Not Found: Facebook," and it says that the location is on [208,9] of the ti:/module.js.  When I scroll down to line 208, I can't identify an error.  This chunk of information is written the same all over the code, but with different content in the text, but it doesn't identify errors with the others. I tried to run the code again in a web browser to see if it would at least display the information, but I was then prompted with this error "TypeError: Undefined is not an object (evaluating 'fb.appid = 975848059099933') at /app.js (line 15)." Again I do not see an error at this location, and I have read through the code and I don't know why the app won't run. Im thinking that titanium doesn't recognize fb for Facebook, although it appears to be defined in the code and there are no errors that jump out at you.  I do know that it isn't something wrong with my emulator, because I just tried running my own project and it works.  But I can't diagnose the problem for why this project won't run for me.