Reviewer 2 (redo)
So I had to remove the code where the facebook function was supposed to run, that is what was giving me problems running the project. The app runs and opens as intended now, and I can successfully navigate between pages. A little too much scrolling for a business card though, and the font could be reduced a little to eliminate this.  The app also has horizontal integration and runs smoothly without words being cut off or overlapping. When I ran it on an iPad emulator it also adjusted to the screen size as intended.  While reading through the code I did not discover any errors and everything seemed to be laid out correctly. Good work!